// src/services/openapi/notification-service/api.ts
export interface GenericBucketRequestDtoNotificationTemplateFilterRequestDto {
    /**
     * 
     * @type {PageRequestDto}
     * @memberof GenericBucketRequestDtoNotificationTemplateFilterRequestDto
     */
    page?: PageRequestDto;
    /**
     * 
     * @type {SortRequestDto}
     * @memberof GenericBucketRequestDtoNotificationTemplateFilterRequestDto
     */
    sortList?: SortRequestDto;
    /**
     * 
     * @type {SearchDetailRequestDto}
     * @memberof GenericBucketRequestDtoNotificationTemplateFilterRequestDto
     */
    searchDetail?: SearchDetailRequestDto;
    /**
     * 
     * @type {NotificationTemplateFilterRequestDto}
     * @memberof GenericBucketRequestDtoNotificationTemplateFilterRequestDto
     */
    filter?: NotificationTemplateFilterRequestDto;
}
// ...
export interface GenericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto {
    /**
     * 
     * @type {PageRequestDto}
     * @memberof GenericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto
     */
    page?: PageRequestDto;
    /**
     * 
     * @type {SortRequestDto}
     * @memberof GenericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto
     */
    sortList?: SortRequestDto;
    /**
     * 
     * @type {SearchDetailRequestDto}
     * @memberof GenericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto
     */
    searchDetail?: SearchDetailRequestDto;
    /**
     * 
     * @type {TransactionNotificationTemplateSubmissionFilterRequestDto}
     * @memberof GenericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto
     */
    filter?: TransactionNotificationTemplateSubmissionFilterRequestDto;
}
// ...
export interface NotificationTemplateFilterRequestDto {
    /**
     * 
     * @type {string}
     * @memberof NotificationTemplateFilterRequestDto
     */
    templateCode?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof NotificationTemplateFilterRequestDto
     */
    templateType?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof NotificationTemplateFilterRequestDto
     */
    language?: string;
    /**
     * 
     * @type {boolean}
     * @memberof NotificationTemplateFilterRequestDto
     */
    isActive?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof NotificationTemplateFilterRequestDto
     */
    isSendEmail?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof NotificationTemplateFilterRequestDto
     */
    isSendLos?: boolean;
    /**
     * 
     * @type {string}
     * @memberof NotificationTemplateFilterRequestDto
     */
    startDate?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationTemplateFilterRequestDto
     */
    endDate?: string;
}
// ...
export interface PageRequestDto {
    /**
     * 
     * @type {number}
     * @memberof PageRequestDto
     */
    noPage?: number;
    /**
     * 
     * @type {number}
     * @memberof PageRequestDto
     */
    itemPerPage?: number;
}
// ...
export interface RequestByIdDtoString {
    /**
     * 
     * @type {string}
     * @memberof RequestByIdDtoString
     */
    id?: string;
}
/**
 * 
 * @export
 * @interface SaveBeforeEditNotificationRequestDto
 */
export interface SaveBeforeEditNotificationRequestDto {
    /**
     * 
     * @type {number}
     * @memberof SaveBeforeEditNotificationRequestDto
     */
    notificationTemplateId?: number;
}
// ...
export interface SearchDetailRequestDto {
    /**
     * 
     * @type {string}
     * @memberof SearchDetailRequestDto
     */
    key?: string;
    /**
     * 
     * @type {string}
     * @memberof SearchDetailRequestDto
     */
    value?: string;
}
/**
 * 
 * @export
 * @interface SortRequestDto
 */
export interface SortRequestDto {
    /**
     * 
     * @type {string}
     * @memberof SortRequestDto
     */
    columnName?: string;
    /**
     * 
     * @type {string}
     * @memberof SortRequestDto
     */
    sortType?: string;
}
// ...
export interface TransactionNotificationTemplateSubmissionFilterRequestDto {
    /**
     * 
     * @type {string}
     * @memberof TransactionNotificationTemplateSubmissionFilterRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {Array<object>}
     * @memberof TransactionNotificationTemplateSubmissionFilterRequestDto
     */
    templateTypes?: Array<object>;
    /**
     * 
     * @type {Array<object>}
     * @memberof TransactionNotificationTemplateSubmissionFilterRequestDto
     */
    statuses?: Array<object>;
    /**
     * 
     * @type {string}
     * @memberof TransactionNotificationTemplateSubmissionFilterRequestDto
     */
    templateCode?: string;
}
// ...
export class NotificationTemplateControllerApi extends BaseAPI {
    /**
     * 
     * @summary API to get detail notification template
     * @param {RequestByIdDtoString} requestByIdDtoString 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof NotificationTemplateControllerApi
     */
    public getDetilNotificationTemplate(requestByIdDtoString: RequestByIdDtoString, options?: any) {
        return NotificationTemplateControllerApiFp(this.configuration).getDetilNotificationTemplate(requestByIdDtoString, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get list notification template
     * @param {GenericBucketRequestDtoNotificationTemplateFilterRequestDto} genericBucketRequestDtoNotificationTemplateFilterRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof NotificationTemplateControllerApi
     */
    public getNotificationTemplate(genericBucketRequestDtoNotificationTemplateFilterRequestDto: GenericBucketRequestDtoNotificationTemplateFilterRequestDto, options?: any) {
        return NotificationTemplateControllerApiFp(this.configuration).getNotificationTemplate(genericBucketRequestDtoNotificationTemplateFilterRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get submission list for approval
     * @param {GenericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto} genericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof NotificationTemplateControllerApi
     */
    public getNotificationTemplateSubmission(genericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto: GenericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto, options?: any) {
        return NotificationTemplateControllerApiFp(this.configuration).getNotificationTemplateSubmission(genericBucketRequestDtoTransactionNotificationTemplateSubmissionFilterRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get submission detail for approval
     * @param {RequestByIdDtoString} requestByIdDtoString 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof NotificationTemplateControllerApi
     */
    public getNotificationTemplateSubmissionDetail(requestByIdDtoString: RequestByIdDtoString, options?: any) {
        return NotificationTemplateControllerApiFp(this.configuration).getNotificationTemplateSubmissionDetail(requestByIdDtoString, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to create transaction from existing notification template
     * @param {SaveBeforeEditNotificationRequestDto} saveBeforeEditNotificationRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof NotificationTemplateControllerApi
     */
    public saveBeforeEditNotificationTemplate(saveBeforeEditNotificationRequestDto: SaveBeforeEditNotificationRequestDto, options?: any) {
        return NotificationTemplateControllerApiFp(this.configuration).saveBeforeEditNotificationTemplate(saveBeforeEditNotificationRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to save transaction notification template
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof NotificationTemplateControllerApi
     */
    public saveTransactionNotificationTemplate(options?: any) {
        return NotificationTemplateControllerApiFp(this.configuration).saveTransactionNotificationTemplate(options).then((request) => request(this.axios, this.basePath));
    }

}